refactor: moved the components to release-manager folder
