+++ static/client/components/ReleaseManager/AddChecksumPanel.tsx
