+++ static/client/components/ReleaseManager/ReleaseField.tsx
@@ -5,7 +5,7 @@ import classNames from "classnames";
 
 import TaggedFieldInput from "./TaggedFieldInput";
 
-import { formatInputLabel, recurseEqual, validateRequiredNumber } from "@/pages/Releases/utils";
+import { formatInputLabel, recurseEqual, validateRequiredNumber } from "./utils";
 import type { ITaggedField, ReleaseFieldValue } from "@/services/api/types/releases";
 import { isTaggedField } from "@/services/api/types/releases";
