+++ static/client/pages/Releases/UpdateChecksumsPage.tsx
@@ -2,16 +2,16 @@ import { type ReactNode, useCallback, useEffect, useState } from "react";
 
 import { useOutletContext } from "react-router-dom";
 
-import type { IReleasesLayoutOutletContext } from "./ReleasesLayout";
-import AddChecksumPanel, { type IChecksumEditTarget } from "./components/AddChecksumPanel";
-import ChecksumCategoryTable from "./components/ChecksumCategoryTable";
+import type { IReleasesOutletContext } from "./Releases";
+import AddChecksumPanel, { type IChecksumEditTarget } from "@/components/ReleaseManager/AddChecksumPanel";
+import ChecksumCategoryTable from "@/components/ReleaseManager/ChecksumCategoryTable";
 
 import { isRecord } from "@/services/api/types/releases";
 import { usePanelsStore } from "@/store/app";
 
 const UpdateChecksumsPage = (): ReactNode => {
   const { formData, handleChecksumAdd, handleChecksumDelete, registerAddChecksum } =
-    useOutletContext<IReleasesLayoutOutletContext>();
+    useOutletContext<IReleasesOutletContext>();
 
   const [addChecksumPanelVisible, toggleAddChecksumPanel] = usePanelsStore((state) => [
     state.addChecksumPanelVisible,
