+++ static/client/pages/Releases/UpdateReleasesPage.tsx
@@ -2,15 +2,15 @@ import { type ReactNode } from "react";
 
 import { useOutletContext } from "react-router-dom";
 
-import type { IReleasesLayoutOutletContext } from "./ReleasesLayout";
-import ReleaseField from "./components/ReleaseField";
-import { formatSectionTitle } from "./utils";
+import type { IReleasesOutletContext } from "./Releases";
+import ReleaseField from "@/components/ReleaseManager/ReleaseField";
+import { formatSectionTitle } from "@/components/ReleaseManager/utils";
 
 import type { ReleaseFieldValue } from "@/services/api/types/releases";
 import { isRecord } from "@/services/api/types/releases";
 
 const UpdateReleasesPage = (): ReactNode => {
-  const { formData, data, handleFieldChange } = useOutletContext<IReleasesLayoutOutletContext>();
+  const { formData, data, handleFieldChange } = useOutletContext<IReleasesOutletContext>();
   const originalReleases = data.releases;
 
   const releaseCategories = Object.entries(formData).filter(([key]) => key !== "checksums");
