+++ static/client/pages/Releases/_ReleaseForm.scss
@@ -10,7 +10,7 @@
     }
 
     input {
-      margin: 0 0 0.4rem 0;
+      margin: 0 0 0.6rem 0;
     }
   }
